complete

+++ src/aws_databridge/main.py
@@ -1,4 +1,4 @@
-from aws_databridge.utils.convert_to import text_to_csv, json_to_csv, xml_to_csv
+from aws_databridge.utils.convert_to import text_to_csv, json_to_csv, xml_to_csv, csv_to_json
 from aws_databridge.db_import.rds import import_to_rds
 from aws_databridge.db_import.dynamodb import import_to_dynamodb
 from aws_databridge.db_import.document import import_to_documentdb
@@ -26,19 +26,25 @@ def main():
     db_choice = input('What database would you like to import this data to? (rds = 1, dynamo = 2, neptune = 3, documentDB = 4): ')
     table_name = input('Enter the name of the table you would like to import this data to: ')
 
+  
     #convert supported non csv types to csv
     if file_type == 'txt':
+         print('Converting data to csv...')
+         file = text_to_csv(file)
+         print('Data has been converted to csv')
+    elif file_type == 'json' and db_choice != '3':
         print('Converting data to csv...')
-        text_to_csv(file)
+        file = json_to_csv(file)
         print('Data has been converted to csv')
-    elif file_type == 'json':
-            print('Converting data to csv...')
-            json_to_csv(file)
-            print('Data has been converted to csv')
     elif file_type == 'xml':
-            print('Converting data to csv...')
-            xml_to_csv(file)
-            print('Data has been converted to csv')
+        print('Converting data to csv...')
+        file = xml_to_csv(file)
+        print('Data has been converted to csv')
+    if db_choice == '3' and file_type == 'csv':
+        print('Converting data to json...')
+        file = csv_to_json(file)
+        print('Data has been converted to json')
+
 
     # import data to the selected database, allowing for multiple imports if neccessary 
     stop = 0;
@@ -49,6 +55,7 @@ def main():
             primary_key = input('Enter the primary key for the table: ')
             import_to_dynamodb(file, table_name, primary_key)
         elif db_choice == '3':
+            import_to_neptune(file)
             print('Data has been imported to Neptune')
         elif db_choice == '4':
             import_to_documentdb(file)
@@ -57,14 +64,16 @@ def main():
         if stop == 0:
             db_choice = input('What database would you like to import this data to? (rds = 1, dynamo = 2, neptune = 3, documentDB = 4): ')
             table_name = input('Enter the name of the table you would like to import this data to: ')
-    # Delete the file after processing
-    try:
-        os.remove(file)
-        print(f"File '{file}' has been deleted from the EC2 instance.")
-    except FileNotFoundError:
-        print(f"File '{file}' not found.")
-    except Exception as e:
-        print(f"Error deleting file: {e}")
+
+    # Delete the S3 file after processing
+    if import_path == '2':
+        try:
+            os.remove(file)
+            print(f"File '{file}' has been deleted from the EC2 instance.")
+        except FileNotFoundError:
+            print(f"File '{file}' not found.")
+        except Exception as e:
+            print(f"Error deleting file: {e}")
 
 if __name__ == '__main__':
     main()
